% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_attribute}
\alias{add_attribute}
\title{#' Add class
#'
#' @param x Object to add a class to
#' @param cls Class to be added
#'
#' @export
#' @return `x` with class `cls` added
#' @examples
add_class <- function(x,cls="mapfile"){
  class(x) <- c(cls,class(x)) %>% unique
  x
}
Add attribute}
\usage{
add_attribute(graf, value, attr)
}
\arguments{
\item{graf}{A causal map}

\item{value}{A value to add as attribute}

\item{attr}{The attribute to add to}
}
\value{
The causal map `graf` but with an additional attribute `attr` with value `value`
}
\description{
#' Add class
#'
#' @param x Object to add a class to
#' @param cls Class to be added
#'
#' @export
#' @return `x` with class `cls` added
#' @examples
add_class <- function(x,cls="mapfile"){
  class(x) <- c(cls,class(x)) %>% unique
  x
}
Add attribute
}
