% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_class}
\alias{add_class}
\title{Add class}
\usage{
add_class(x, cls = "mapfile")
}
\arguments{
\item{x}{Object to add a class to}

\item{cls}{Class to be added}
}
\value{
`x` with class `cls` added
}
\description{
Add class
}
