% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assemble_map}
\alias{assemble_map}
\title{Assemble map}
\usage{
assemble_map(
  factors = NULL,
  links = NULL,
  statements = NULL,
  sources = NULL,
  questions = NULL,
  settings = NULL
)
}
\arguments{
\item{settings}{}
}
\value{

}
\description{
Assemble map
}
