% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coerce_mapfile}
\alias{coerce_mapfile}
\title{Create mapfile}
\usage{
coerce_mapfile(tables)
}
\arguments{
\item{tables}{we assume that the initial map has standard format}
}
\value{
A clean map in which all issues have been resolved.
}
\description{
Create mapfile
}
