% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_map}
\alias{create_map}
\title{Create map}
\usage{
create_map(
  factors = NULL,
  links = NULL,
  statements = NULL,
  sources = NULL,
  questions = NULL,
  settings = NULL,
  clean = T
)
}
\arguments{
\item{clean}{}
}
\value{
A tidy map. If both factors and links are NULL, and empty map is returned.
}
\description{
Create map
}
