% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_flows}
\alias{get_flows}
\title{Get flow}
\usage{
get_flows(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{

}
\description{
Uses maximum flow / minimum cut algorithm
}
