% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_mapfile_from_s3}
\alias{get_mapfile_from_s3}
\title{Get causal map candidate from s3}
\usage{
get_mapfile_from_s3(path)
}
\arguments{
\item{path}{A path to an s3 file.}
}
\value{
The s3 file. Not necessarily a valid causal map.
}
\description{
Get causal map candidate from s3
}
