% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_in_hierarchy}
\alias{is_in_hierarchy}
\title{Update map}
\usage{
is_in_hierarchy(labels)
}
\arguments{
\item{clean}{}
}
\value{
A tidy map. If both factors and links are NULL, and empty map is returned.
}
\description{
Update map
}
