% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_map}
\alias{load_map}
\title{Load map}
\usage{
load_map(
  path = NULL,
  factors = NULL,
  links = NULL,
  statements = NULL,
  sources = NULL,
  questions = NULL,
  settings = NULL,
  clean = T,
  connection = conn
)
}
\arguments{
\item{clean}{}
}
\value{
A tidy map. If both factors and links are NULL, and empty map is returned.
}
\description{
Load map
}
