% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_igraph}
\alias{make_igraph}
\title{Make igraph (from factors and links)}
\usage{
make_igraph(factors, links, use_labels = F)
}
\arguments{
\item{use_labels}{}
}
\value{

}
\description{
Make igraph (from factors and links)
}
