% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_igraph_from_links}
\alias{make_igraph_from_links}
\title{Clean map}
\usage{
make_igraph_from_links(links)
}
\arguments{
\item{settings}{}
}
\value{
A clean map in which all issues have been resolved.
}
\description{
Clean map
}
