% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_interactive_map}
\alias{make_interactive_map}
\title{Make a visNetwork}
\usage{
make_interactive_map(graf, scale = 1, safe_limit = 200)
}
\arguments{
\item{graf}{A tidymap. The factors table and links table may contain additional formatting information like color.background.}

\item{scale}{Increase from the default 1 to make the map taller.}
}
\value{
A visnetwork
}
\description{
Make a visNetwork (https://datastorm-open.github.io/visNetwork/) from a tidymap.
}
