% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_map_metrics}
\alias{make_map_metrics}
\title{Make map-level metrics}
\usage{
make_map_metrics(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{field}{A numerical field in the factor table which will control the scale.}
}
\value{
A tidymap with a new or overwritten column `size`in the factor table varying between .2 and 1.
}
\description{
Make map-level metrics
}
