% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_print_map}
\alias{make_print_map}
\title{Make a Graphviz map}
\usage{
make_print_map(
  graf = NULL,
  maxwidth = NULL,
  grv_layout = NULL,
  grv_splines = NULL,
  grv_overlap = NULL,
  color = NULL,
  rainbow = F,
  graph_title = "",
  ranksep_slider = NULL,
  nodesep_slider = NULL,
  safe_limit = NULL
)
}
\arguments{
\item{graf}{A mapfile. Link and factor tables may contain columns to control formatting
such as `color.border`.}

\item{grv_layout}{What layout to use. Default is `dot`.}

\item{grv_splines}{How to create splines. See Graphviz documentation.}

\item{grv_overlap}{See Graphviz documentation.}

\item{color}{Default font color}

\item{safe_limit}{Integer. Large maps with many edges can take a long time to layout.
If !is.null(safe_limit), the resulting map is simplified by bundling edges and selecting
most frequent factors.}
}
\description{
Make a Graphviz map: https://graphviz.org/documentation/
}
