% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_table}
\alias{make_table}
\title{Make table}
\usage{
make_table(graf, tab)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{tab}{factors links or statements}
}
\value{
The corresponding table
}
\description{
Make table
}
