% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_vn}
\alias{make_vn}
\title{Make a visnetwork}
\usage{
make_vn(graf, scale = 1)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{scale}{}
}
\value{
A visnetwork
}
\description{
Make a visnetwork
}
