% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_map}
\alias{merge_map}
\title{Merge map}
\usage{
merge_map(map1, map2)
}
\value{
A tidy map. The column *_map_id is set to reflect the id of the map.
}
\description{
This also has a wrapper, pipe_merge_map
}
