% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_commands}
\alias{parse_commands}
\title{Parse commands}
\usage{
parse_commands(graf, tex)
}
\arguments{
\item{graf}{A tidygraph}

\item{tex}{A set of commands to parse, separated by linebreaks if there is more than one command.
Each line consists of two words corresponding to the name of the pipe function to be applied, e.g. `color links` calls the function `color_links`
The function name is followed by field=value pairs corresponding to the arguments of the function such as `top=10`.}
}
\value{
A tidygraph, the result of successively applying the commands to the input graph.
}
\description{
A parser which breaks a text input into individual commands.
There is one command format corresponding to each of
the family of pipe functions.
}
\examples{
graf \%>\% parse_commands("select factors top=10\\nbundle links field=n\\ncolor links field=n") \%>\% make_vn
}
