% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pip_fix_columns}
\alias{pip_fix_columns}
\title{Fix columns}
\usage{
pip_fix_columns(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A tidymap with a additional columns.
}
\description{
Fix columns
}
