% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_clean_map}
\alias{pipe_clean_map}
\title{Clean map}
\usage{
pipe_clean_map(tables = NULL)
}
\arguments{
\item{settings}{}
}
\value{
A clean map in which all issues have been resolved.
}
\description{
Clean map
}
