% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_cluster_factors}
\alias{pipe_cluster_factors}
\title{Cluster factors}
\usage{
pipe_cluster_factors(graf, clusters = NULL)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{

}
\description{
Cluster factors
}
