% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_coerce_mapfile}
\alias{pipe_coerce_mapfile}
\title{Coerce mapfile}
\usage{
pipe_coerce_mapfile(tables)
}
\arguments{
\item{tables}{we assume that the initial map has standard format}
}
\value{
A clean map in which all issues have been resolved.
}
\description{
Coerce mapfile
}
