% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_color_links}
\alias{pipe_color_links}
\title{Color links}
\usage{
pipe_color_links(
  graf,
  field = "frequency",
  lo = "#FCFDBF",
  hi = "#5F187F",
  mid = "#D3436E",
  fixed = NULL,
  fun = "length"
)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{field}{A numerical or character field in the link table.
If it is character, the other parameters are ignored and a color-blind friendly qualitative palette is applied.}

\item{lo}{Optionally, a color specification for the low end of the color range. Default is `green`.}

\item{hi}{Optionally, a color specification for the high end of the color range. Default is `blue`.}

\item{mid}{Optionally, a color specification for the middle of the color range. Default is `gray`.}

\item{fixed}{Optionally, a color specification which will be applied everywhere and overrides `field`.}
}
\value{
A tidymap with a new or overwritten column `color`in the link table.
}
\description{
Color links
}
