% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_combine_opposites}
\alias{pipe_combine_opposites}
\title{pipe_combine_opposites}
\usage{
pipe_combine_opposites(graf, flipchar = "~", add_colors = T)
}
\arguments{
\item{add_colors}{Applies only to links}
}
\description{
pipe_combine_opposites
}
