% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_compact_factors_links}
\alias{pipe_compact_factors_links}
\title{Clean map}
\usage{
pipe_compact_factors_links(mapfile)
}
\arguments{
\item{settings}{}
}
\value{
A clean map in which all issues have been resolved.
}
\description{
Clean map
}
