% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_dash_links}
\alias{pipe_dash_links}
\title{Dash links}
\usage{
pipe_dash_links(
  graf,
  field = "hashtags",
  operator = "notequals",
  value = "",
  type = "dashed"
)
}
\arguments{
\item{type}{}
}
\description{
Dash links
}
