% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_fade_factors}
\alias{pipe_fade_factors}
\title{Fade factors}
\usage{
pipe_fade_factors(graf, field = "n")
}
\arguments{
\item{graf}{A tidygraph}

\item{field}{A numerical field in the link table which will control the amount of fading
(the alpha value of the factors).}
}
\value{
A tidygraph in which the column `color.background`in the factor table has alpha proportionate to the values in `field`.
}
\description{
Fade factors
}
