% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_fade_links}
\alias{pipe_fade_links}
\title{Fade links}
\usage{
pipe_fade_links(graf, field = "n")
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{field}{A numerical field in the link table which will control the amount of fading
(the alpha value of the links).}
}
\value{
A tidymap in which the column `color`in the link table has alpha proportionate to the values in `field`.
}
\description{
Fade links
}
