% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_find_factors}
\alias{pipe_find_factors}
\title{Find factors}
\usage{
pipe_find_factors(graf, field = NULL, value, operator = NULL, up = 0, down = 0)
}
\arguments{
\item{field}{Field (column, variable) to search}

\item{value}{Value to search for}

\item{operator}{c('contains','notcontains','=','notequals','greater','less','starts','ends').
How to search.}

\item{up}{integer. Default is 0.}

\item{down}{integer. Default is 0.}
}
\value{
A tidymap containing only matching factors; if `up`!=0 then also factors this number of steps
upstream of the matching factors are also included and likewise for `down`!=0.
The links are filtered correspondingly to return only the "ego network" i.e. links between the returned factors.

If operator and field are both NULL, the value is treated as a simple search string for the field `label`.
}
\description{
When field==label, you can search for whole labels and also for including keywords or flags which are common to more than one factor.
Fields may be from the original data and/or fields created by the app, e.g. `n` (frequency).
When field is 'label', 'value' can contain a vector of search terms separated by ' OR '.
}
\examples{
pipe_find_factors(CashTransferMap,"Cash")
pipe_find_factors(CashTransferMap,field="label",value="Cash",operator="contains")
}
