% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_find_links}
\alias{pipe_find_links}
\title{Find links}
\usage{
pipe_find_links(graf, field, value, operator = "=")
}
\arguments{
\item{operator}{c('contains','notcontains','=','notequals','greater','less','starts','ends')}
}
\value{

}
\description{
Find links
}
