% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_find_statements}
\alias{pipe_find_statements}
\title{Find statements}
\usage{
pipe_find_statements(graf, field, value, operator = "=")
}
\arguments{
\item{field}{Field (column, variable) to search}

\item{value}{Value to search for}

\item{operator}{c('contains','notcontains','=','notequals','greater','less','starts','ends').
How to search.}
}
\value{

}
\description{
Find statements
}
