% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_fix_columns}
\alias{pipe_fix_columns}
\title{Fix columns}
\usage{
pipe_fix_columns(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A tidymap with a additional columns.
}
\description{
Fix columns
}
