% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_hide_factors}
\alias{pipe_hide_factors}
\title{Hide factors}
\usage{
pipe_hide_factors(graf, value)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{value}{The string to find within factor labels}
}
\value{
A map in which labels matching the string are excluded
}
\description{
Hide factors
}
