% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_label_factors}
\alias{pipe_label_factors}
\title{Label factors}
\usage{
pipe_label_factors(graf, field = "n", clear = F)
}
\arguments{
\item{graf}{A tidygraph}

\item{field}{A numerical or character field in the factor table.}

\item{clear}{Logical. Whether to clear any existing labels or to concatenate the new result after
any existing labels. Default is `FALSE`.}
}
\value{
A tidygraph with a column `label`. If `clear` is FALSE, the new label is concatenated
after any existing label. The new label is of the form `field: value`.
}
\description{
Label factors
}
