% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_label_links}
\alias{pipe_label_links}
\title{Label links}
\usage{
pipe_label_links(graf, field = "n", clear = F)
}
\arguments{
\item{graf}{A tidygraph representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{field}{A numerical or character field in the link table.}

\item{clear}{Logical. Whether to clear any existing labels or to concatenate the new result after
any existing labels.}
}
\value{
A tidygraph with a column `label`. If `clear` is FALSE (the default), the new label is concatenated
after any existing label. The new label is of the form `field: value`.
}
\description{
Label links
}
