% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_merge_map}
\alias{pipe_merge_map}
\title{Merge map}
\usage{
pipe_merge_map(graf, path)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{clean}{}
}
\value{
A tidy map. If both factors and links are NULL, and empty map is returned.
}
\description{
Merge map
}
