% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_merge_mapfile}
\alias{pipe_merge_mapfile}
\title{Pipe merge map}
\usage{
pipe_merge_mapfile(graf, path)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A tidy map. The column *_map_id is set to reflect the id of the map.
}
\description{
A wrapper around merge_mapfile to make it work in the app.
}
