% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_merge_statements}
\alias{pipe_merge_statements}
\title{Merge statements into links}
\usage{
pipe_merge_statements(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A tidymap in which columns from the statements table are merged into the links table.
}
\description{
Merge statements into links
}
\examples{
cashTransferMap \%>\% pipe_merge_statements() \%>\% pipe_find_links(field="text",value="women",operator="contains")
cashTransferMap \%>\% pipe_find_statements(field="text",value="women",operator="contains")
}
