% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_normalise_factors_links}
\alias{pipe_normalise_factors_links}
\title{#' Fix columns
#'
#' @inheritParams parse_commands
#'
#' @return A tidymap with a additional columns.
#' @export
#'
#'
#' @examples
pipe_fix_columns <- function(graf){
  graf %>% update_map(
    factors = graf$factors %>% fix_columns_factors(),
    links = graf$links %>% fix_columns_links()
  )
}
Normalise factors and links}
\usage{
pipe_normalise_factors_links(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A tidymap with a additional columns.
}
\description{
#' Fix columns
#'
#' @inheritParams parse_commands
#'
#' @return A tidymap with a additional columns.
#' @export
#'
#'
#' @examples
pipe_fix_columns <- function(graf){
  graf %>% update_map(
    factors = graf$factors %>% fix_columns_factors(),
    links = graf$links %>% fix_columns_links()
  )
}
Normalise factors and links
}
