% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_recalculate_all}
\alias{pipe_recalculate_all}
\title{Fix columns}
\usage{
pipe_recalculate_all(graf)
}
\arguments{
\item{graf}{A mapfile representing a causal map.
A mapfile is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A mapfile with a additional columns.
}
\description{
Fix columns
}
