% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_reconstruct_factors_from_links}
\alias{pipe_reconstruct_factors_from_links}
\title{Pipe reconstruct factors from links}
\usage{
pipe_reconstruct_factors_from_links(graf)
}
\arguments{
\item{graf}{}
}
\description{
Pipe reconstruct factors from links
}
