% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_remove_isolated}
\alias{pipe_remove_isolated}
\title{Remove isolated factors}
\usage{
pipe_remove_isolated(graf)
}
\arguments{
\item{graf}{}
}
\value{

}
\description{
Removes any factors which have no links.
This can be useful after any function like pipe_select_links() which remove links.
}
