% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_remove_orphaned_links}
\alias{pipe_remove_orphaned_links}
\title{Remove orphaned links}
\usage{
pipe_remove_orphaned_links(graf)
}
\arguments{
\item{graf}{A mapfile representing a causal map.
A mapfile is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A mapfile with a additional columns.
}
\description{
Remove orphaned links
}
