% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_remove_statementless_links}
\alias{pipe_remove_statementless_links}
\title{Pipe remove statementless links}
\usage{
pipe_remove_statementless_links(graf)
}
\arguments{
\item{graf}{}
}
\description{
Pipe remove statementless links
}
