% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_scale_links}
\alias{pipe_scale_links}
\title{Scale factors}
\usage{
pipe_scale_links(graf, field = "frequency", fixed = NULL, fun = "length")
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{field}{A numerical field in the link table which will control the scale (the width of the links).}
}
\value{
A tidymap with a new or overwritten column `width`in the link table varying between .2 and 1.
}
\description{
Scale factors
}
