% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_search_factors}
\alias{pipe_search_factors}
\title{Search factors}
\usage{
pipe_search_factors(graf, value, operator = "contains", up = 0, down = 0, ...)
}
\arguments{
\item{operator}{c('contains','notcontains','=','notequals','greater','less','starts','ends')}

\item{up}{integer}

\item{down}{integer}
}
\value{

}
\description{
A wrapper around pipe_find factors, providing a simpler syntax
}
\details{
Wrapper for pipe_find_factors
}
