% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_select_links}
\alias{pipe_select_links}
\title{Select links}
\usage{
pipe_select_links(graf, top = NULL, bottom = NULL)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{top}{Bundle the links and select only the `top` links in terms of their frequency}

\item{bottom}{Bundle the links and select only the `bottom` links in terms of their frequency}

\item{is_proportion}{}
}
\value{

}
\description{
Select links
}
