% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_trace_continuity}
\alias{pipe_trace_continuity}
\title{Pipe trace continuity}
\usage{
pipe_trace_continuity(graf, field = "source_id")
}
\arguments{
\item{graf}{A mapfile representing a causal map.
A mapfile is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A mapfile.
}
\description{
Pipe trace continuity
}
