% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_trace_robustness}
\alias{pipe_trace_robustness}
\title{Trace robustness}
\usage{
pipe_trace_robustness(graf, from, to, length = 4, field = NULL)
}
\arguments{
\item{graf}{A mapfile representing a causal map.
A mapfile is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{field}{An optional field by which to split the calculate_robustness calculation and output.}
}
\value{

}
\description{
This is a powerful command which allows the user to trace paths from one or more upstream factors to one or more downstream factors.
Only links which are part of such paths are displayed.
This function wraps trace_paths and calculate_robustness
}
