% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_trace_threads}
\alias{pipe_trace_threads}
\title{Pipe trace threads}
\usage{
pipe_trace_threads(graf, field = "source_id", direction = "down")
}
\arguments{
\item{graf}{A mapfile resulting from pipe_trace_paths or pipe_trace_robustness}
}
\value{
A mapfile with additional links fields contined_down_sid, n_unique_from_before_continued, n_unique_from_before
}
\description{
A different approach from robustness: gets added on after a trace paths filter.
}
