% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_wrap_factors}
\alias{pipe_wrap_factors}
\title{Wrap factor labels}
\usage{
pipe_wrap_factors(graf, length = 20)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{length}{line length}
}
\value{
A tidymap with factor labels wrapped to `length`
}
\description{
Wrap factor labels
}
