% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_wrap_links}
\alias{pipe_wrap_links}
\title{Wrap link labels}
\usage{
pipe_wrap_links(graf, length = 20)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{length}{line length}
}
\value{
A tidymap with link labels wrapped to `length`
}
\description{
Wrap link labels
}
