% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_zoom_factors}
\alias{pipe_zoom_factors}
\title{Zoom factors}
\usage{
pipe_zoom_factors(
  graf,
  level = 1,
  separator = ";",
  preserve_frequency = +Inf,
  frequency_field = "frequency",
  hide = T
)
}
\arguments{
\item{graf}{A mapfile representing a causal map.
A mapfile is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{

}
\description{
Another important ability of the Causal Map Viewer is to manipulate maps which use
hierarchical coding, which is a very powerful way to code causal information.
It is possible to globally by roll up all factors into up to a certain level, for example,
to roll them all up to their top level.
But it is also possible to selectively zoom individual hierarchies while leaving others intact to any given level.
}
\details{
Zoom out from a map, merging factors within the saem hierarchy
}
