% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_visual_bundles}
\alias{prepare_visual_bundles}
\title{Prepare visual bundles}
\usage{
prepare_visual_bundles(
  graf,
  group = NULL,
  color_field = NULL,
  color_fun = NULL,
  size_field = NULL,
  size_fun = NULL,
  label_field = NULL,
  label_fun = NULL,
  lo = "",
  mid = "",
  hi = ""
)
}
\arguments{
\item{hi}{}
}
\value{

}
\description{
Prepare visual bundles
}
