% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_join}
\alias{update_join}
\title{#' Fix columns
#'
#' @inheritParams parse_commands
#'
#' @return A mapfile with a additional columns.
#' @export
#'
#'
#' @examples
pipe_fix_columns <- function(graf){
  graf %>% update_mapfile(
    factors = graf$factors %>% fix_columns_factors(),
    links = graf$links %>% fix_columns_links()
  )
}}
\usage{
update_join(old, new)
}
\description{
#' Fix columns
#'
#' @inheritParams parse_commands
#'
#' @return A mapfile with a additional columns.
#' @export
#'
#'
#' @examples
pipe_fix_columns <- function(graf){
  graf %>% update_mapfile(
    factors = graf$factors %>% fix_columns_factors(),
    links = graf$links %>% fix_columns_links()
  )
}
}
