% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_map}
\alias{update_map}
\title{Update map}
\usage{
update_map(
  map,
  factors = NULL,
  links = NULL,
  statements = NULL,
  sources = NULL,
  questions = NULL,
  settings = NULL,
  tables = list(factors = NULL, links = NULL, statements = NULL, sources = NULL,
    questions = NULL, settings = NULL),
  clean = T,
  all = T
)
}
\arguments{
\item{clean}{}
}
\value{
A tidy map. If both factors and links are NULL, and empty map is returned.
}
\description{
Update map
}
