% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_mapfile}
\alias{update_mapfile}
\title{Update mapfile}
\usage{
update_mapfile(
  map,
  tables = NULL,
  factors = NULL,
  links = NULL,
  statements = NULL,
  sources = NULL,
  questions = NULL,
  settings = NULL,
  clean = T,
  all = T
)
}
\arguments{
\item{clean}{}
}
\value{
A tidy map. If both factors and links are NULL, and empty map is returned.
}
\description{
Update mapfile
}
