% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_attribute}
\alias{add_attribute}
\title{Add attribute}
\usage{
add_attribute(graf, value)
}
\arguments{
\item{graf}{A mapfile}

\item{value}{A value to add as attribute}

\item{attr}{The attribute to add to}
}
\value{
The mapfile `graf` but with an additional attribute `attr` with value `value`
}
\description{
Add attribute
}
