% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assemble_map}
\alias{assemble_map}
\title{Title}
\usage{
assemble_map(
  factors = NULL,
  links = NULL,
  statements = NULL,
  sources = NULL,
  questions = NULL,
  settings = NULL,
  tables = NULL
)
}
\arguments{
\item{settings}{}
}
\value{
A clean map in which all issues have been resolved.
}
\description{
Title
}
