% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{assemble_mapfile}
\alias{assemble_mapfile}
\title{Assemble map}
\usage{
assemble_mapfile(
  factors = standard_factors(),
  links = standard_links(),
  statements = standard_statements(),
  sources = standard_sources(),
  questions = standard_questions(),
  settings = standard_settings()
)
}
\description{
The only difference between this and just creating a list is that it provides names
}
