% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_mapfile_from_excel}
\alias{get_mapfile_from_excel}
\title{Title}
\usage{
get_mapfile_from_excel(path)
}
\arguments{
\item{path}{The path of an .xslx file to load. The .xlsx file can have worksheets called `statements`, `links`, `factors`, `sources` and `questions`.}
}
\description{
Title
}
