% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_grviz}
\alias{make_grviz}
\title{Make a Graphviz map}
\usage{
make_grviz(
  graf = NULL,
  maxwidth = NULL,
  grv_cluster = F,
  grv_cluster_select = NULL,
  grv_layout = "dot",
  grv_splines = "splines",
  grv_overlap = F,
  color = "grey",
  ranksep_slider = 3,
  nodesep_slider = 20,
  wrap_slider = 12,
  safe_limit = T
)
}
\arguments{
\item{safe_limit}{}
}
\value{

}
\description{
Make a Graphviz map
}
