% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_map_metrics}
\alias{make_map_metrics}
\title{Make map-level metrics}
\usage{
make_map_metrics(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A tibble with map-level metrics
}
\description{
Make map-level metrics
}
