% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_vn}
\alias{make_vn}
\title{Make a visnetwork}
\usage{
make_vn(graf, scale = 1)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{scale}{}
}
\value{
A visnetwork
}
\description{
Make a visnetwork
}
