% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_line}
\alias{parse_line}
\title{Parse line}
\usage{
parse_line(line1, graf)
}
\arguments{
\item{graf}{A mapfile representing a causal map.}

\item{line}{A line of text to be parsed}
}
\value{
A list containing the function name and a list of parameters.
}
\description{
The engine for parse_commands
}
