% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pip_metrics}
\alias{pip_metrics}
\title{Create factor metrics}
\usage{
pip_metrics(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}
}
\value{
A tidymap with new or overwritten columns for factor metrics.
}
\description{
Create factor metrics
}
