% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_bundle_factors}
\alias{pipe_bundle_factors}
\title{Bundle factors}
\usage{
pipe_bundle_factors(graf, value = "")
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{value}{A search string.}
}
\value{
A tidymap in which factors matching the search string are merged into one, with rerouted links.
If the search string is empty, factors with the same first word are grouped.
}
\description{
Bundle factors
}
