% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_bundle_links}
\alias{pipe_bundle_links}
\title{Bundle links}
\usage{
pipe_bundle_links(graf, counter = "n", group = NULL)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{field}{}
}
\value{
A tidymap in which sets of coterminal, same-direction links are replaced with
one link (when `field` = 'n') or more than one link for each of the values of `field`
present in the data. In each case, each new link has a field n representing the number
of links it is replacing, unless the links it is replacing already had values n in which
case the new value of `n` is the sum of the `n` values of the constituent links.
}
\description{
Bundle links
}
