% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_clean_map}
\alias{pipe_clean_map}
\title{Clean map}
\usage{
pipe_clean_map(graf)
}
\arguments{
\item{graf}{}
}
\value{
A cleaned map done by taking it apart and putting it back together
}
\description{
Clean map
}
