% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_fade_links}
\alias{pipe_fade_links}
\title{Fade links}
\usage{
pipe_fade_links(graf, field = "n")
}
\arguments{
\item{graf}{A tidygraph}

\item{field}{A numerical field in the link table which will control the amount of fading
(the alpha value of the links).}
}
\value{
A tidygraph in which the column `color`in the link table has alpha pr\oportionate to the values in `field`.
}
\description{
Fade links
}
