% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_find_factors}
\alias{pipe_find_factors}
\title{Find factors}
\usage{
pipe_find_factors(graf, field, value, operator = "=", up = 0, down = 0)
}
\arguments{
\item{down}{}
}
\value{

}
\description{
Find factors
}
