% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_find_links}
\alias{pipe_find_links}
\title{Find links}
\usage{
pipe_find_links(graf, field = NULL, value, operator = NULL)
}
\description{
Find links
}
\examples{
pipe_find_links(cashTransferMap,value="Cash")
pipe_find_links(cashTransferMap,field="label",value="Cash",operator="contains")
pipe_find_links(cashTransferMap,field="from",value="12",operator="greater")
}
