% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_find_links}
\alias{pipe_find_links}
\title{Find links}
\usage{
pipe_find_links(graf, field = NULL, value, operator = NULL)
}
\arguments{
\item{operator}{c('contains','notcontains','=','notequals','greater','less','starts','ends')}
}
\value{

}
\description{
If operator and field are both NULL, the value is treated as a simple search string for the field `label`.
}
\examples{
pipe_find_links(CashTransferMap,"Cash")
pipe_find_links(CashTransferMap,field="label",value="Cash",operator="contains")
pipe_find_links(CashTransferMap,field="from",value="12",operator="greater")
}
