% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_label_factors}
\alias{pipe_label_factors}
\title{Label factors}
\usage{
pipe_label_factors(graf, field = "n", clear = F)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called `factors` and edges are called `links.`}

\item{field}{A numerical or character field in the factor table.}

\item{clear}{Logical. Whether to clear any existing labels or to concatenate the new result after
any existing labels. Default is `FALSE`.}
}
\value{
A tidymap with a column `label`. If `clear` is FALSE, the new label is concatenated
after any existing label. The new label is of the form `field: value`.
}
\description{
For example it is possible to add the factor frequency to the factor labels.
 More than one label can be added: Labels are additive and are applied one after the other.
 Other formatting commands such as factor and link colour can either specify a fixed colour (`links colour blue`)
 or can be used conditionally, so that a the colour or transparency of links can depend on any custom field existing in the original data
 such as, say, `gender` but also on fields created by the app e.g. frequency.
}
