% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_label_links}
\alias{pipe_label_links}
\title{Label links}
\usage{
pipe_label_links(
  graf,
  field = "frequency",
  fun = "unique",
  clear = T,
  field_label = F
)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{field}{A numerical or character field in the link table.}

\item{clear}{Logical. Whether to clear any existing labels or to concatenate the new result after
any existing labels.}
}
\value{
A tidymap with a column `label`. If `clear` is FALSE (the default), the new label is concatenated
after any existing label. The new label is of the form `field: value`.
}
\description{
Label links
}
