% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_recode_factors}
\alias{pipe_recode_factors}
\title{Pipe recode factors}
\usage{
pipe_recode_factors(
  graf,
  recodes = graf$settings$recodes,
  top_level_only = T,
  retain_unmatched = T
)
}
\arguments{
\item{recodes}{}
}
\description{
Pipe recode factors
}
