% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_remove_brackets}
\alias{pipe_remove_brackets}
\title{Remove bracketed expressions}
\usage{
pipe_remove_brackets(graf, value = "[")
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{length}{line length}
}
\value{
A tidymap with factor labels wrapped to `length`
}
\description{
Remove bracketed expressions
}
