% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_scale_factors}
\alias{pipe_scale_factors}
\title{Scale factors}
\usage{
pipe_scale_factors(graf, field = "n")
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{field}{A numerical field in the factor table which will control the scale.}
}
\value{
A tidymap with a new or overwritten column `size`in the factor table varying between .2 and 1.
}
\description{
Scale factors
}
