% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_search_factors}
\alias{pipe_search_factors}
\title{Search factors
Wrapper for pipe_find_factors}
\usage{
pipe_search_factors(graf, value, operator = "contains", up = 0, down = 0, ...)
}
\arguments{
\item{down}{}
}
\value{

}
\description{
Search factors
Wrapper for pipe_find_factors
}
