% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_select_factors}
\alias{pipe_select_factors}
\title{Select factors}
\usage{
pipe_select_factors(
  graf,
  top = 10,
  bottom = NULL,
  all = F,
  field = "frequency"
)
}
\arguments{
\item{graf}{A mapfile representing a causal map.
A mapfile is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{top}{Select only the `top` factors in terms of their frequency}

\item{is_proportion}{}
}
\description{
Select factors
}
