% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_select_links}
\alias{pipe_select_links}
\title{Select links}
\usage{
pipe_select_links(graf, top = NULL, bottom = NULL)
}
\arguments{
\item{graf}{A mapfile representing a causal map.
A mapfile is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{top}{Bundle the links and select only the `top` links in terms of their frequency}

\item{bottom}{Bundle the links and select only the `bottom` links in terms of their frequency}

\item{is_proportion}{}
}
\value{

}
\description{
Select links
}
\examples{
select links top=20 should keep all the links within the 20 fattest bundles/sets of links with the most links in each bundle.
But previously the algorithm actually combined them into 20 individual links (i.e. create radically fewer actual links) and just remembered the frequency. Now, it keeps the individual links (so a map with select links top =3 might still have 3000 actual links if there were 1000 from A to B and 1000 from B to C and 1000 from C to D. By default, the Interactive and Print maps would indeed combine these into three thick pipes for performance sake, but there would still be 3000 links there somewhere.
}
