% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_select_links}
\alias{pipe_select_links}
\title{Select links}
\usage{
pipe_select_links(graf, top)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called `factors` and edges are called `links.`}

\item{top}{Bundle the links and select only the `top` links in terms of their frequency}

\item{is_proportion}{}
}
\value{

}
\description{
Select links
}
