% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_trace_continuity}
\alias{pipe_trace_continuity}
\title{Pipe trace continuity}
\usage{
pipe_trace_continuity(graf, field = "source_id")
}
\arguments{
\item{graf}{A mapfile resulting from pipe_trace_paths or pipe_trace_robustness}
}
\value{
A mapfile with additional links fields continuation_id, n_unique_incoming_continued, n_unique_incoming
}
\description{
A different approach from robustness: gets added on after a trace paths filter.
}
