% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_trace_paths}
\alias{pipe_trace_paths}
\title{Trace paths}
\usage{
pipe_trace_paths(graf, from, to, length)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{length}{}
}
\value{

}
\description{
Trace paths
}
