% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_trace_paths}
\alias{pipe_trace_paths}
\title{Trace paths}
\usage{
pipe_trace_paths(graf, from, to, length = 4)
}
\arguments{
\item{graf}{A mapfile representing a causal map.
A mapfile is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{

}
\description{
This is a powerful command which allows the user to trace paths from one or more upstream factors to one or more downstream factors.
Only links which are part of such paths are displayed.
}
