% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_wrap_factors}
\alias{pipe_wrap_factors}
\title{Wrap factor labels}
\usage{
pipe_wrap_factors(graf, length = 20)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}

\item{length}{line length}
}
\value{
A tidymap with factor labels wrapped to `length`
}
\description{
Wrap factor labels
}
