% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipe_zoom_factors}
\alias{pipe_zoom_factors}
\title{Zoom factors
Zoom out from a map, merging factors within the saem hierarchy}
\usage{
pipe_zoom_factors(graf, level, separator, hide)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called factors and edges are called links.}

\item{hide}{}
}
\value{

}
\description{
Zoom factors
Zoom out from a map, merging factors within the saem hierarchy
}
