% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tibbles}
\alias{tibbles}
\alias{factors_table}
\alias{links_table}
\alias{statements_table}
\alias{sources_table}
\alias{questions_table}
\alias{settings_table}
\alias{links_table_full}
\title{Extracting tibbles from A tidymap}
\usage{
factors_table(graf)

links_table(graf)

statements_table(graf)

sources_table(graf)

questions_table(graf)

settings_table(graf)

links_table_full(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes,
with an optional additional table of statements.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A tibble.
}
\description{
These three functions extract tables of factors, links or statements from a tidymap.
}
