% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tibbles}
\alias{tibbles}
\alias{factor_table}
\alias{link_table}
\title{Extracting tibbles from a tidygraph}
\usage{
factor_table(graf)

link_table(graf)
}
\arguments{
\item{graf}{A tidygraph}
}
\value{
A tibble.
}
\description{
Extracting tibbles from a tidygraph
}
