% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tibbles}
\alias{tibbles}
\alias{factors_table}
\alias{links_table}
\alias{statements_table}
\title{Extracting tibbles from A tidymap}
\usage{
factors_table(graf)

links_table(graf)

statements_table(graf)
}
\arguments{
\item{graf}{A tidymap representing a causal map.
In this package, nodes are called `factors` and edges are called `links.`}
}
\value{
A tibble.
}
\description{
Extracting tibbles from A tidymap
}
